geo-rep / gsyncd: rework internal RPC event model
So far the concurrency model adopted by the RePCe RPC backed was,
in theory, M:N -- that is, requests coming from whatever number
of client side contexts were handled by a thread pool of unrelated
size on the slave side.

This had the following issues:

- in practice, M == 1, ie. there was just a single thread that
  synchronously makes RePCe requests (ie. there is no use of a
  multithreaded server)
- due to the Python threading implementation¹, the thread pool
  of the server was an overhead, not a gain
- also, server was not thread-safe, as with heavy load + large
  responses, the response messages could end up interleaved
- We had one tunable for gsyncd concurrency, "sync-jobs"; this
  determined both the size of the rsync pool on master side and
  size of RePCe server thread pool on slave side. Indeed, these
  two resources are not related, so contolling them by the same
  value is an artificial constraint (that came to existence due
  to the lack of a better idea).

Now we move to a 1:1 model, and the actual number of contexts is
2 (can be forced to be 1).

Master creates two gsyncd slaves, a "foreground" and a "background"
one. Basic crawl procedure makes a linear progress that sends
synchronous requests -- these are sent to the foreground slave. However,
there are a lot of requests for which we either don't need to wait at
all (eg. set_xtime) or at least not at the point of making the request
(eg. setattr -- the continuation of the crawl does not depend
on setattr response, it's enough to ensure that a setattr failure
vetoes set_xtime). Now these are sended to the background server,
asynchronously.

- refactored RePCe to have single threaded server and to make
  async messages eaiser
- refactored master to send async requests wherever feasible
- refactored resources (in particular, SSH) to support "clones"
- refactored gsyncd.c wrapper to allow rsync invocations that
  have multiple gsyncd siblings
- new tunable: "max-slaves" for controlling number of slave
  processes; if it's 1, "foreground" and "background" slaves
  will be the same. Now "sync-jobs" controls only the size of the
  rsync worker pool, and it defaults to the number of cores
  (cf. sysconf(3)).

These changes are compatible over the wire with release-3.2
gsynd; however earlier 3.3 betas (that feature the gsyncd.c
wrapper) will support only max-slaves=1 mode of operation.

¹ http://www.dabeaz.com/python/GIL.pdf

Change-Id: Id5a9927e8ec2179d5b33663b3f05bb0482d54242
Signed-off-by: Csaba Henk <csaba@redhat.com>