geo-rep / gsyncd: gsyncd binary wrapper context fixes
Changes to __gf_calloc (commit ed4b76ba) triggers a recursive
sequence of _gf_log() and __glusterfs_this_location() calls. This
is due to __gf_calloc() trying to access THIS which is not valid
in the context of geo replication binary wrapper. This patch fixes
the problem by introducing the relevant context in the wrapper.

Change-Id: I32b0b6927074692dbf66bad2f3e93434201aaa82
BUG: 849304
Signed-off-by: Venky Shankar <vshankar@redhat.com>